+++ internal/component/otelcol/receiver/receiver_test.go
@@ -161,9 +161,12 @@ func newTestEnvironment(t *testing.T, onTracesConsumer func(t otelconsumer.Trace
 		},
 	}
 
+	ctrl, err := componenttest.NewControllerFromReg(util.TestLogger(t), reg)
+	require.NoError(t, err)
+
 	return &testEnvironment{
 		t:          t,
-		Controller: componenttest.NewControllerFromReg(util.TestLogger(t), reg),
+		Controller: ctrl,
 	}
 }